Add Azure/EntraID route for SAML
Adds support for SAML via the optional `azure` extra.
This can be configured with the `config.auth.azure` options.
