+++ goosebit/db/__init__.py
@@ -1,3 +1,4 @@
+import asyncio
 from logging import getLogger
 
 from tortoise import Tortoise
@@ -9,13 +10,14 @@ from goosebit.db.models import Device
 logger = getLogger(__name__)
 
 
-async def init() -> bool:
+async def init():
     await Tortoise.init(config=TORTOISE_CONF)
     try:
         await Device.first()
     except OperationalError:
-        return False
-    return True
+        logger.exception("DB does not exist, try running `poetry run aerich upgrade`.")
+        loop = asyncio.get_running_loop()
+        loop.stop()
 
 
 async def close():
