+++ pyproject.toml
@@ -26,9 +26,11 @@ httpx = "^0.27.0"
 pydantic-settings = "^2.4.0"
 
 asyncpg = { version = "^0.29.0", optional = true }
+msal = { version = "^1.31.0", optional = true }
 
 [tool.poetry.extras]
 postgresql = ["asyncpg"]
+azure = ["msal"]
 
 [tool.poetry.group.dev.dependencies]
 isort = "^5.13.2"
